+++ packages/google_maps_flutter/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapController.java
@@ -279,7 +279,7 @@ final class GoogleMapController
           }
           break;
         }
-      case "map#takeScreenshot":
+      case "map#takeSnapshot":
         {
           if (googleMap != null) {
             final MethodChannel.Result _result = result;